[Budget] Fix consumed_amount calculation from feature allocations
Root cause: consumed_amount was hardcoded to 0 in all 3 product budget
endpoints. Schema also used int type instead of float.

Fix:
- Added _calculate_consumed_amount() helper function that traverses:
  product_budget → budget_lines → FeatureBudgetLineAllocation
  → RoadmapFeature.total_cost_keur * allocation_percentage
- Applied to get_product_budgets, create_or_update_product_budget,
  get_product_budget_detail
- Changed ProductBudgetResponse.consumed_amount, remaining_amount,
  allocated_amount from int to float to support fractional kEUR values

Verified:
- Baggage Reconciliation System: consumed=19.85 kEUR ✅
- Utilization = 0.58% (19.85 / 3423) ✅
- Zero-allocation products show 0.0 consumed ✅
- No divide-by-zero errors ✅

Modules affected: Phase 2 Budget
Risk: 🟡 Medium

+++ backend/app/schemas/budget_config.py
@@ -101,9 +101,9 @@ class ProductBudgetResponse(BaseModel):
     id: UUID
     budget_version_id: UUID
     product: ProductInfo
-    allocated_amount: int
-    consumed_amount: int = 0
-    remaining_amount: int = 0
+    allocated_amount: float
+    consumed_amount: float = 0.0
+    remaining_amount: float = 0.0
     utilization_percentage: float = 0.0
     budget_lines_count: int = 0
     created_at: datetime
